fix(reporting): TAMOC-381: Consolidated facility-server and central-server initReporting logic in the database package
