+++ packages/central-server/app/database/index.js
@@ -1,2 +1,1 @@
-export  { initDatabase, closeDatabase } from './connection';
-export { initReporting } from './reporting';
+export { initDatabase, closeDatabase } from './connection';
