+++ packages/facility-server/__tests__/utilities.js
@@ -16,9 +16,9 @@ import { Problem } from '@tamanu/errors';
 import { ReadSettings } from '@tamanu/settings';
 import { chance } from '@tamanu/fake-data/fake';
 import { asNewRole, showError } from '@tamanu/shared/test-helpers';
+import { initReporting } from '@tamanu/database/services/reporting';
 
 import { createApiApp } from '../dist/createApiApp';
-import { initReporting } from '../dist/database';
 import { buildToken } from '../dist/middleware/auth';
 
 import { toMatchTabularReport } from './toMatchTabularReport';
